[MODIFY] fun add in CommandController is done

+++ src/main/kotlin/Coordinates.kt
@@ -1,4 +1,3 @@
-class Coordinates (var x: Float, var y: Double) {
-//    private var x: Float = TODO()
-//    private var y: Float = TODO()
+class Coordinates (var x: Float, var y: Float) {
+
 }
